requirements.txt
arrow==0.12.1
backports.functools-lru-cache==1.5
bintrees==2.0.7
CacheControl==0.12.5
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
click==6.7
dateparser==0.7.0
django-apptemplates==1.4 
django-papertrail==1.1.9
Django==1.11.15
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.8.2
enum34==1.1.6
grpcio
google-cloud-firestore==0.29.0
gunicorn==19.6.0
firebase-admin==2.13.0
futures==3.2.0
google-api-core==1.4.0
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-storage==1.12.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
idna==2.7
Markdown==2.6.11
moment==0.8.2
msgpack==0.5.6
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-dateutil==2.7.3
pytz==2018.5
redis==2.10.6
regex==2018.8.29
requests==2.19.1
rq==0.12.0
rsa==3.4.2
six==1.11.0
times==0.7
tzlocal==1.5.1
urllib3==1.23
